requirements.txt
click==8.0.3
psycopg2-binary
python-dotenv==0.19.2
SQLAlchemy==1.4.31
